reorganized. tests are failing but because of the de-coupling, not the reorg
