+++ src/sinks/datadog/traces/mod.rs
@@ -11,11 +11,11 @@ mod integration_tests;
 #[cfg(test)]
 mod tests;
 
+pub(crate) mod apm_stats;
 mod config;
 mod request_builder;
 mod service;
 mod sink;
-mod stats;
 
 #[allow(warnings, clippy::pedantic, clippy::nursery)]
 pub(crate) mod ddsketch_full {
@@ -28,5 +28,3 @@ pub(crate) mod dd_proto {
 }
 
 pub use self::config::DatadogTracesConfig;
-pub use self::request_builder::build_request;
-pub use self::stats::Aggregator;
